broke out edit and delete middleware for projects, renamed some files to make more sense
