+++ api/middleware/validateAuthData.js
@@ -6,13 +6,11 @@ const Projects = require('../models/projects-model');
 const validateNewUser = async (req, res, next) => {
 	const { email, username, password, role } = req.body;
 
-	// form field validation
 	if (!email) return res.status(400).json({ message: 'Email required' });
 	if (!username) return res.status(400).json({ message: 'Username required' });
 	if (!password) return res.status(400).json({ message: 'Password required' });
 	if (!role) return res.status(400).json({ message: 'Role required' });
 
-	// check for uniqueness
 	try {
 		const [checkEmail, checkUsername] = await Promise.all([
 			Users.findByEmail(email.toLowerCase()),
@@ -38,35 +36,7 @@ const validateLogin = (req, res, next) => {
 	next();
 };
 
-const validateEditProject = async (req, res, next) => {
-	try {
-		const project = await Projects.findProjectById(req.params.id);
-		const { uid, isAdmin, superUser } = req.token;
-
-		if (!req.token) return res.status(400).json({ message: 'Could not find credentials' });
-
-		// make sure user trying to edit ticket is either an admin, superUser or author
-		return uid === project.stakeholder || uid === project.project_manager || superUser || isAdmin
-			? next()
-			: res.status(400).json({ message: 'Sorry, you do not have permission to edit this ticket' });
-	} catch (err) {
-		console.log(err);
-	}
+module.exports = {
+	validateNewUser,
+	validateLogin
 };
-
-const validateDeleteProject = async (req, res, next) => {
-	try {
-		const project = await Projects.findProjectById(req.params.id);
-		const { uid, isAdmin, superUser } = req.token;
-
-		if (!req.token) return res.status(400).json({ message: 'Could not find credentials' });
-
-		return uid === project.stakeholder || uid === project.project_manager || superUser || isAdmin
-			? next()
-			: res.status(400).json({ errMessage: 'Sorry, you do not have permission to delete this project' });
-	} catch (err) {
-		console.log(err);
-	}
-};
-
-module.exports = { validateNewUser, validateLogin, validateEditProject, validateDeleteProject };
