+++ api/routes/projects-router.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
 
-const restrictUsers = require('../middleware/validateUsers');
+const { validateEditProject, validateDeleteProject } = require('../middleware/validateProjectWriting');
 
 const Projects = require('../models/projects-model');
 
@@ -41,7 +41,7 @@ router.post('/', async (req, res) => {
 });
 
 // PUT - edit project
-router.put('/:id', restrictUsers.validateEditProject, async (req, res) => {
+router.put('/:id', validateEditProject, async (req, res) => {
 	try {
 		const project = await Projects.editProject(req.params.id, req.body, req.token);
 
@@ -53,7 +53,7 @@ router.put('/:id', restrictUsers.validateEditProject, async (req, res) => {
 });
 
 // DELETE - delete project
-router.delete('/:id', restrictUsers.validateDeleteProject, async (req, res) => {
+router.delete('/:id', validateDeleteProject, async (req, res) => {
 	try {
 		const deleted = await Projects.deleteProject(req.params.id, req.token);
